refactor #258: Rename test directory local_kv_client to embedded_client
Align test directory naming with EmbeddedClient type.

Changes:
- Rename d-engine-server/tests/local_kv_client/ → embedded_client/
- Rename local_kv_client_operations_embedded.rs → embedded_client_operations.rs
- Update module declarations in mod.rs and integration_test.rs

Benefits:
- Consistent naming with EmbeddedClient API
- Clearer intent (embedded mode client tests)
