+++ d-engine-server/tests/integration_test.rs
@@ -5,9 +5,9 @@ mod cluster_membership;
 mod cluster_state_and_metadata;
 mod common;
 mod consistent_reads;
+mod embedded_client;
 mod failover_and_recovery;
 mod leader_election;
-mod local_kv_client;
 mod replication_and_sync;
 
 // Storage layer integration tests
